module-1/studio/requirements.txt
langgraph
langgraph-prebuilt
langchain-core
langchain-community
langchain-openai


langchain-aws
boto3
python-dotenv
